✨ Highlight Search Term inside child rows
Responsive Datatables will move columns that don't fit into a table down into a child row. This row can be shown/hidden via toggles that are added to the datatable.
However, `datatables.mark.js` does not mark terms inside these rows. By applying the mark function to the tableBody we can mark terms inside the body, but outside of the table columns.

First we mark the tableBody, when looping through the nodes, we unmark anything that was previously marked to avoid a double mark.